services/worker/app/requirements.txt
fastapi
uvicorn[standard]
google-cloud-storage
google-auth
pydantic